+++ server/src/main/java/com/linecorp/centraldogma/server/internal/api/auth/RequiresAdministratorDecorator.java
@@ -27,8 +27,8 @@ import com.linecorp.armeria.server.SimpleDecoratingService;
 import com.linecorp.armeria.server.annotation.Decorator;
 import com.linecorp.armeria.server.annotation.DecoratorFactoryFunction;
 import com.linecorp.centraldogma.server.internal.admin.auth.AuthUtil;
-import com.linecorp.centraldogma.server.internal.admin.auth.User;
 import com.linecorp.centraldogma.server.internal.api.HttpApiUtil;
+import com.linecorp.centraldogma.server.metadata.User;
 
 /**
  * A {@link Decorator} to allow a request from an administrator only.
