+++ server/src/main/java/com/linecorp/centraldogma/server/metadata/MigrationUtil.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 LINE Corporation
+ * Copyright 2019 LINE Corporation
  *
  * LINE Corporation licenses this file to you under the Apache License,
  * version 2.0 (the "License"); you may not use this file except in compliance
@@ -14,13 +14,13 @@
  * under the License.
  */
 
-package com.linecorp.centraldogma.server.internal.metadata;
+package com.linecorp.centraldogma.server.metadata;
 
 import static com.linecorp.centraldogma.server.command.Command.createRepository;
-import static com.linecorp.centraldogma.server.internal.metadata.MetadataService.METADATA_JSON;
-import static com.linecorp.centraldogma.server.internal.metadata.MetadataService.TOKEN_JSON;
-import static com.linecorp.centraldogma.server.internal.metadata.RepositorySupport.convertWithJackson;
 import static com.linecorp.centraldogma.server.internal.storage.project.ProjectInitializer.INTERNAL_PROJ;
+import static com.linecorp.centraldogma.server.metadata.MetadataService.METADATA_JSON;
+import static com.linecorp.centraldogma.server.metadata.MetadataService.TOKEN_JSON;
+import static com.linecorp.centraldogma.server.metadata.RepositorySupport.convertWithJackson;
 import static java.util.Objects.requireNonNull;
 import static java.util.stream.Collectors.toMap;
