+++ server/src/main/java/com/linecorp/centraldogma/server/metadata/package-info.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 LINE Corporation
+ * Copyright 2019 LINE Corporation
  *
  * LINE Corporation licenses this file to you under the Apache License,
  * version 2.0 (the "License"); you may not use this file except in compliance
@@ -18,6 +18,6 @@
  * Metadata management for Central Dogma projects.
  */
 @NonNullByDefault
-package com.linecorp.centraldogma.server.internal.metadata;
+package com.linecorp.centraldogma.server.metadata;
 
 import com.linecorp.centraldogma.common.util.NonNullByDefault;
