+++ server/src/test/java/com/linecorp/centraldogma/server/metadata/MetadataServiceTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 LINE Corporation
+ * Copyright 2019 LINE Corporation
  *
  * LINE Corporation licenses this file to you under the Apache License,
  * version 2.0 (the "License"); you may not use this file except in compliance
@@ -14,11 +14,11 @@
  * under the License.
  */
 
-package com.linecorp.centraldogma.server.internal.metadata;
+package com.linecorp.centraldogma.server.metadata;
 
-import static com.linecorp.centraldogma.server.internal.metadata.PerRolePermissions.NO_PERMISSION;
-import static com.linecorp.centraldogma.server.internal.metadata.PerRolePermissions.READ_ONLY;
-import static com.linecorp.centraldogma.server.internal.metadata.PerRolePermissions.READ_WRITE;
+import static com.linecorp.centraldogma.server.metadata.PerRolePermissions.NO_PERMISSION;
+import static com.linecorp.centraldogma.server.metadata.PerRolePermissions.READ_ONLY;
+import static com.linecorp.centraldogma.server.metadata.PerRolePermissions.READ_WRITE;
 import static com.linecorp.centraldogma.server.storage.project.Project.REPO_META;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -31,7 +31,6 @@ import com.linecorp.centraldogma.common.ChangeConflictException;
 import com.linecorp.centraldogma.common.ProjectExistsException;
 import com.linecorp.centraldogma.common.RepositoryExistsException;
 import com.linecorp.centraldogma.server.command.Command;
-import com.linecorp.centraldogma.server.internal.admin.auth.User;
 import com.linecorp.centraldogma.testing.internal.ProjectManagerRule;
 
 public class MetadataServiceTest {
